Remove py_sleep test
I think this was added for historical reasons when upstream plpython
was not interrumpible, but that's not the case anymore.

Right now, either we adapt it to use plpython2u or plpython3u depending
on what's available, or we just remove it since upstream was fixed
several years ago. It's simpler to remove it